Lint update

+++ source/components/debugger/dbfileio.c
@@ -2,7 +2,7 @@
  *
  * Module Name: dbfileio - Debugger file I/O commands.  These can't usually
  *              be used when running the debugger in Ring 0 (Kernel mode)
- *              $Revision: 1.82 $
+ *              $Revision: 1.83 $
  *
  ******************************************************************************/
 
@@ -316,7 +316,7 @@ AcpiDbReadTable (
 
 
     fseek (fp, 0, SEEK_END);
-    FileSize = ftell (fp);
+    FileSize = (UINT32) ftell (fp);
     fseek (fp, 0, SEEK_SET);
 
     /* Read the table header */
